requirements.txt
Flask
requests
psutil
python-dotenv
Pillow
